Restructure fieldValidators and test so we have a separate file for different validators - and we also needed to avoid circular dependencies.
