+++ src/lib/fieldValidator.ts
@@ -1,0 +1,42 @@
+export interface IFieldValidator {
+  validate(value: string): IFieldValidationResponse;
+}
+
+export interface IFieldValidationResponse {
+  errorMessages: string[];
+  valid: boolean;
+  invalid: boolean;
+}
+
+export interface FieldRule {
+  errorMessage: string;
+  predicateFn: (valueToValidate: string) => boolean;
+}
+
+export abstract class FieldValidator implements IFieldValidator {
+  protected _rules: FieldRule[];
+
+  validate(value: string): IFieldValidationResponse {
+    return {
+      valid: this.isValid(value),
+      invalid: !this.isValid(value),
+      errorMessages: this.errorMessages(value),
+    };
+  }
+
+  static valueViolatesRule(value: string, rule: FieldRule): boolean {
+    return rule.predicateFn(value);
+  }
+
+  private isValid(value: string): boolean {
+    return !this._rules.some((rule) =>
+      FieldValidator.valueViolatesRule(value, rule)
+    );
+  }
+
+  private errorMessages(value: string): string[] {
+    return this._rules
+      .filter((rule) => FieldValidator.valueViolatesRule(value, rule))
+      .map((rule) => rule.errorMessage);
+  }
+}
